test: move issue 64822 regression into existing suite
