+++ pkg/expression/integration_test/integration_test.go
@@ -4458,3 +4458,37 @@ func TestDeepCopyRetType(t *testing.T) {
 	tk.MustExec("create view v0(c0) as select cast((t1.c0 div t1.c0) as decimal) from t1;")
 	tk.MustQuery("select * from v0 inner join t0 on (v0.c0 like cast(v0.c0 as char) <= t0.c0) and (not atan2(t0.c0, v0.c0));").Check(testkit.Rows())
 }
+
+func TestIssue64822LikePredicateOrderIndependence(t *testing.T) {
+	testWithVecMode := func(t *testing.T, vecMode string) {
+		store := testkit.CreateMockStore(t)
+		tk := testkit.NewTestKit(t, store)
+		tk.MustExec("use test")
+		tk.MustExec("set tidb_enable_vectorized_expression = " + vecMode)
+		tk.MustExec("set tidb_opt_projection_push_down = 0")
+		tk.MustExec("set tidb_hash_join_version = 'legacy'")
+		tk.MustExec("drop table if exists t0, t1")
+		tk.MustExec("create table t0(c0 char, c1 blob)")
+		tk.MustExec("create table t1 like t0")
+		tk.MustExec("insert into t0 values ('_', '1')")
+		tk.MustExec("insert into t1 values ('ᡖ', '0')")
+		tk.MustQuery(
+			"select t1.c0 like t0.c0, t1.c0 not like t0.c0, not(case true when false then t1.c1 else t1.c0 end) from t1 join t0",
+		).Check(testkit.Rows("1 0 1"))
+		tk.MustQuery(
+			"select t1.c0 like t0.c0, not(case true when false then t1.c1 else t1.c0 end), t1.c0 not like t0.c0 from t1 join t0",
+		).Check(testkit.Rows("1 1 0"))
+		tk.MustQuery(
+			"select * from t1 join t0 where t1.c0 like t0.c0 and t1.c0 not like t0.c0 and not(case true when false then t1.c1 else t1.c0 end)",
+		).Check(testkit.Rows())
+		tk.MustQuery(
+			"select * from t1 join t0 where t1.c0 like t0.c0 and not(case true when false then t1.c1 else t1.c0 end) and t1.c0 not like t0.c0",
+		).Check(testkit.Rows())
+	}
+	t.Run("vec_on", func(t *testing.T) {
+		testWithVecMode(t, "on")
+	})
+	t.Run("vec_off", func(t *testing.T) {
+		testWithVecMode(t, "off")
+	})
+}
